config-schema: break out schema browser and scroll targets context
